siegewar - rename for compat. with town ruin pr
