+++ src/com/palmergames/bukkit/towny/tasks/HourlyTimerTask.java
@@ -2,9 +2,8 @@ package com.palmergames.bukkit.towny.tasks;
 
 import com.palmergames.bukkit.towny.Towny;
 import com.palmergames.bukkit.towny.TownySettings;
-import com.palmergames.bukkit.towny.utils.PostRespawnPeacefulnessUtil;
 import com.palmergames.bukkit.towny.utils.TownPeacefulnessUtil;
-import com.palmergames.bukkit.towny.war.common.ruins.RuinsUtil;
+import com.palmergames.bukkit.towny.war.common.townruin.TownRuinUtil;
 import com.palmergames.bukkit.towny.war.siegewar.SiegeWarTimerTaskController;
 
 /**
@@ -23,7 +22,7 @@ public class HourlyTimerTask extends TownyTimerTask {
 	@Override
 	public void run() {
 		if (TownySettings.getWarCommonTownRuinsEnabled()) {
-			RuinsUtil.evaluateRuinedTownRemovals();
+			TownRuinUtil.evaluateRuinedTownRemovals();
 		}
 
 		if(TownySettings.getWarCommonPeacefulTownsEnabled()) {
